+++ templates/index.html
@@ -197,14 +197,18 @@
 				<div class="features-grid">
 					<i class="fa fa-cogs" aria-hidden="true"></i> 属性设置
 				</div>
-				<div id="steps" class="col-sm-4 col-md-4 settings"></div>
-				<div id="options" class="col-sm-8 col-md-8 settings"></div>
-				<div class="clearfix"></div>
 				<div class="stats-agileinfo agileits-w3layouts">
+					<div id="steps" class="col-sm-4 col-md-4 settings"></div>
+					<div id="options" class="col-sm-8 col-md-8 settings"></div>
+					<div class="clearfix"></div>
 					<div class="col-md-12">
 						<hr>
 						<button id="RunButton" type="button" class="btn btn-primary btn-lg">开始</button>
 					</div>
+				</div>
+				<div class="col-md-12">
+					<br>
+					<br>
 				</div>
 			</div>
 		</div>
@@ -223,8 +227,7 @@
 	<!-- js -->
 	<script src="{{vendorsJs}}"></script>
 	<script src="{{commonJs}}"></script>
-	<script src="static/js/progress.js"></script>
-	<script src="static/js/steps.js"></script>
+	<script src="static/js/main.js"></script>
 	<script>
 		$(document).ready(function () {
 			const outputImg = $('#outputImg'), downImg = $('#downImg')
@@ -233,7 +236,6 @@
 			const app = window.app
 			const progress = app.setup({
 				onProgress,
-				beforeSend: app.submit,
 				success: (result, progress) => {
 					progress.final(app.texts.finish)
 					let im = app.getResource(result)
@@ -241,12 +243,10 @@
 					downImg.attr('href', im)
 				},
 				path,
-				progress: $('#progress'),
 				inputImg: $('#inputImg'),
-				dropZone: $('#drop_zone')
+				dropZone: $('#drop_zone'),
+				features: ['input', 'output', 'SR', 'DN', 'dehaze', 'ednoise', 'resize']
 			})
-			app.initListeners()
-			app.context.setFeatures(['SR', 'DN', 'ednoise', 'dehaze', 'resize', 'input'])
 		})
 	</script>
 </body>
